Grade 7 fix batch B: chat retries and cancel, Yahoo sign-in and injury codes, Keychain-safe engine, test hygiene
Chat: two retries with backoff on 429, 529 and 5xx, and 529 reads as
overloaded; a question survives the panel closing or the screen
switching and its answer keeps arriving; Cancel aborts an answer and
keeps the partial text marked as cut short; the desktop spend moves for
phone-asked turns; a phone question during a desktop turn is refused
with a reason. The in-flight claim registry is owned by AppState, not a
process-global, so two servers in one process never see each other's
claims.

Yahoo: the token pair is written to the Keychain only when it changed,
not every tick; the Settings row re-asks status when the dialog closes
and on sign-out; a sign-in abandoned mid-way frees its loopback
listener (new yahoo_cancel_connect, wired from the dialog); the loopback
accepts in a loop so a speculative browser connection cannot swallow
the redirect; percent decoding is byte-safe; an empty refresh token is
a sign-out. Yahoo injury codes reach the shared classifier and camp PUP
is priced as early weeks, not the season, before kickoff.

Backend: Engine takes its SecretStore at construction, so every test
built engine and YahooState is sandboxed by construction and a scratch
config can no longer migrate a key into the real Keychain; api_key and
store_api_key go through that store; a source scan guards it. A
headshot miss is recorded only on 404 or 410. Cache write failures are
logged once per file.

Tests: poll-until helpers replace fixed sleeps in the companion, TLS
keeper and Yahoo wire tests; the device Item is passed into
CompanionServer instead of a process-global; vitest coverage thresholds
are real numbers; CI reuses the vite build for Playwright; the
logged-coverage scanner reads nested Err patterns correctly; the port
test cannot overflow near the top of the range. Over-cap files split:
App.tsx, commands_chat_tests.rs, yahoo_auth_wire.

+++ draft-assistant/src-tauri/src/companion/net.rs
@@ -420,7 +420,11 @@ mod tests {
         let (held, port) = bind_from(0).expect("an ephemeral port is free");
         let (next, taken) = bind_from(port).expect("the next port is free");
         assert_ne!(taken, port, "the held port was handed out twice");
-        assert!(taken > port && taken < port + super::PORT_ATTEMPTS);
+        // u32 so a kernel-picked port near the top of the range cannot
+        // overflow the bound.
+        assert!(
+            taken > port && u32::from(taken) < u32::from(port) + u32::from(super::PORT_ATTEMPTS)
+        );
         drop(next);
         drop(held);
     }
